Move ApiError next to ApiResult
